rf- merge ft-routine branch  to work with topnavigator implemented in ft-food branch
