+++ src/navigator/BottomNavigator.tsx
@@ -1,14 +1,14 @@
 import React from 'react';
 import { createMaterialBottomTabNavigator } from '@react-navigation/material-bottom-tabs';
+
 import { useThemeContext } from '../theme';
 import Screens from '../screens';
 import RegimenIcon from '../../assets/icons/regimen';
 import HairCareIcon from '../../assets/icons/hair_care';
 import SettingIcon from '../../assets/icons/settings';
 import TabBarLabel from '../commons/tab-bar-label';
-import { IconContainer } from './styles';
 import HomeTopNavigator from './HomeTopNavigator';
-import RoutineScreen from '../screens/regimen/RegimenScreen';
+import { IconContainer } from './styles';
 
 const Tab = createMaterialBottomTabNavigator();
 
@@ -26,8 +26,6 @@ export default function BottomNavigator() {
       <Tab.Screen
         name="HomeTopNavigator"
         component={HomeTopNavigator}
-        name="RegimenScreen"
-        component={Screens.RegimenScreen}
         options={{
           tabBarIcon: ({ color, focused }) => (
             <IconContainer>
