+++ src/screens/regimen/food/styles.ts
@@ -1,16 +1,5 @@
 import styled from 'styled-components/native';
-import applyScale from '../../utils/applyScale';
-
-export const Container = styled.View`
-  flex: 1;
-`;
-
-export const Welcome = styled.Text`
-  font-size: ${({ theme }) => theme.fonts.LARGE_SIZE}px;
-  font-family: ${({ theme }) => theme.fonts.JOST_MEDIUM};
-  color: ${({ theme }) => theme.colors.FONT_DARK_COLOR};
-  text-transform: capitalize;
-`;
+import applyScale from '../../../utils/applyScale';
 
 export const HeaderTitleContainer = styled.View`
   flex-direction: row;
@@ -98,10 +87,4 @@ export const DescriptionContainer = styled.View`
   flex: 1;
   flex-direction: row;
   flex-wrap: wrap;
-`;
-
-export const ButtonContainer = styled.View`
-  flex: 1;
-  justify-content: space-between;
-  margin: 20px;
 `;
