+++ src/screens/regimen/index.tsx
@@ -1,20 +1,20 @@
-import React, { useRef } from 'react';
+import React from 'react';
+import { StatusBar } from 'react-native';
+
 import { NavigationInterface } from '../types';
-import { StatusBar, TouchableOpacity } from 'react-native';
 import { useThemeContext } from '../../theme';
 import SafeAreaView from '../../commons/safe-area-view';
 import Header from '../../commons/header';
 import Button from '../../components/button';
-import SubHeaderSection from './sub_header_section';
+
 import {
   Container,
   Welcome,
   HeaderTitleContainer,
   HeaderTitle,
   HeaderTitleLabel,
-  HeaderTab
+  ButtonContainer
 } from './styles';
-import applyScale from '../../utils/applyScale';
 
 interface RegimenScreenProp extends NavigationInterface {
   testID?: string;
@@ -23,31 +23,6 @@ interface RegimenScreenProp extends NavigationInterface {
 export default function RegimenScreen(props: RegimenScreenProp) {
   const { colors } = useThemeContext();
   const { navigation } = props;
-  const headerTabRef = useRef({
-    routineRef: null,
-    foodRef: null,
-    saloonVisitsRef: null
-  });
-
-  const handleClick = (ref: any, routePath: string) => {
-    //First set existing textDecorationLine to none
-    Object.keys(headerTabRef.current).forEach(key =>
-      headerTabRef.current[key].setNativeProps({
-        style: {
-          textDecorationLine: 'none',
-          opacity: 0.5
-        }
-      })
-    );
-
-    //Now set textDecorationLine to underline for this ref
-    ref.setNativeProps({
-      style: {
-        textDecorationLine: 'underline',
-        opacity: 1
-      }
-    });
-  };
 
   return (
     <SafeAreaView>
@@ -63,60 +38,19 @@ export default function RegimenScreen(props: RegimenScreenProp) {
           </HeaderTitleContainer>
         )}
       />
-      <HeaderTab>
-        <TouchableOpacity
-          onPress={() => {
-            handleClick(headerTabRef.current.routineRef, '');
-          }}
-        >
-          <HeaderTitleLabel
-            ref={ref => {
-              headerTabRef.current.routineRef = ref;
-            }}
-            style={{ opacity: 0.5 }}
-          >
-            routine
-          </HeaderTitleLabel>
-        </TouchableOpacity>
-        <TouchableOpacity
-          onPress={() => {
-            handleClick(headerTabRef.current.foodRef, '');
-          }}
-        >
-          <HeaderTitleLabel
-            style={{
-              marginLeft: applyScale(30),
-              marginRight: applyScale(30),
-              opacity: 0.5
-            }}
-            ref={ref => {
-              headerTabRef.current.foodRef = ref;
-            }}
-          >
-            foods
-          </HeaderTitleLabel>
-        </TouchableOpacity>
-        <TouchableOpacity
-          onPress={() => {
-            handleClick(headerTabRef.current.saloonVisitsRef, '');
-          }}
-        >
-          <HeaderTitleLabel
-            ref={ref => {
-              headerTabRef.current.saloonVisitsRef = ref;
-            }}
-            style={{ opacity: 0.5 }}
-          >
-            saloon visits
-          </HeaderTitleLabel>
-        </TouchableOpacity>
-      </HeaderTab>
       <Container>
-        <Button
-          title="Start Regimen Setup"
-          onPress={() => navigation.navigate('RegimenSetupScreen')}
-        />
         <Welcome>Regimen Screen</Welcome>
+
+        <ButtonContainer>
+          <Button
+            title="Start Regimen Setup"
+            onPress={() => navigation.navigate('RegimenSetupScreen')}
+          />
+          <Button
+            title="See Activities"
+            onPress={() => navigation.navigate('ActivityScreen')}
+          />
+        </ButtonContainer>
       </Container>
     </SafeAreaView>
   );
