+++ src/screens/regimen/routine/style.ts
@@ -1,6 +1,26 @@
 import styled from 'styled-components/native';
-import applyScale from '../../../../utils/applyScale';
-import { Colors } from 'react-native/Libraries/NewAppScreen';
+import applyScale from '../../../utils/applyScale';
+
+export const Container = styled.View`
+  flex: 1;
+  padding: 10px;
+  align-items: center;
+`;
+
+export const Title = styled.Text`
+  text-align: center;
+  margin-top: 20px;
+  margin-bottom: 40px;
+  font-family: ${({ theme: { fonts } }) => fonts.JOST_BOOK};
+  font-size: ${applyScale(30)}px;
+`;
+
+export const SubTitle = styled.Text`
+  text-align: center;
+  margin-top: 18px;
+  opacity: 0.4;
+  font-size: ${applyScale(10)}px;
+`;
 
 export const CardContainer = styled.TouchableOpacity`
   flex: 1;
@@ -51,4 +71,17 @@ export const Content = styled.Text`
   width: ${applyScale(209)}px;
   font-size: ${applyScale(13)}px;
   font-family: ${({ theme }) => theme.fonts.CORMORANT_REGULAR};
+`;
+
+export const HeaderSection = styled.View`
+  flex: 1;
+  align-items: center;
+`;
+
+export const Description = styled.Text`
+  font-size: ${({ theme }) => theme.fonts.LARGE_SIZE * 2}px;
+  font-family: ${({ theme }) => theme.fonts.CORMORANT_REGULAR};
+  color: ${({ theme }) => theme.colors.FONT_DARK_COLOR};
+  margin-top: 7px;
+  align-self: center;
 `;
